refactor: move perp integration test actions to v1 directory
