+++ x/perp/keeper/v1/open_gas_test.go
@@ -13,7 +13,7 @@ import (
 	"github.com/NibiruChain/nibiru/x/common/testutil/assertion"
 	. "github.com/NibiruChain/nibiru/x/oracle/integration/action"
 	perpammtypes "github.com/NibiruChain/nibiru/x/perp/amm/types"
-	. "github.com/NibiruChain/nibiru/x/perp/integration/action"
+	. "github.com/NibiruChain/nibiru/x/perp/integration/action/v1"
 )
 
 func TestOpenGasConsumed(t *testing.T) {
